Separate nearleyc & parser tests
